Move CGAlgorithms functions into Orientation class JTS 4c4ccd7fe5c758abb48647da60dacf8e4cd7f9df

+++ tests/xmltester/tinyxml2/tinyxml2.cpp
@@ -572,7 +572,7 @@ void XMLUtil::ToStr( bool v, char* buffer, int bufferSize )
 */
 void XMLUtil::ToStr( float v, char* buffer, int bufferSize )
 {
-    TIXML_SNPRINTF( buffer, bufferSize, "%.8f", v );
+    TIXML_SNPRINTF( buffer, bufferSize, "%.8g", (double)v );
 }
